+++ Core.UnitTests/CacheControllerUtilTests.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using System.Threading;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using PubComp.Caching.AopCaching;
+using PubComp.Caching.Core.Exceptions;
 
 namespace PubComp.Caching.Core.UnitTests
 {
@@ -81,14 +82,14 @@ namespace PubComp.Caching.Core.UnitTests
             CollectionAssert.AreEquivalent(new string[0], cacheNames);
         }
 
-        [TestMethod][ExpectedException(typeof(CacheException))]
+        [TestMethod][ExpectedException(typeof(CacheClearException))]
         public void TestCacheRegisterCacheItems_NullCacheName()
         {
             this.cacheControllerUtil.GetRegisteredCacheItemKeys(null);
         }
 
         [TestMethod]
-        [ExpectedException(typeof(CacheException))]
+        [ExpectedException(typeof(CacheClearException))]
         public void TestCacheRegisterCacheItems_EmptyCacheName()
         {
             this.cacheControllerUtil.GetRegisteredCacheItemKeys(string.Empty);
@@ -101,14 +102,14 @@ namespace PubComp.Caching.Core.UnitTests
         }
 
         [TestMethod]
-        [ExpectedException(typeof(CacheException))]
+        [ExpectedException(typeof(CacheClearException))]
         public void TestClearCache_FallbackToWildcard()
         {
             this.cacheControllerUtil.ClearCache("cache4");
         }
 
         [TestMethod]
-        [ExpectedException(typeof(CacheException))]
+        [ExpectedException(typeof(CacheClearException))]
         public void TestClearCache_NotFound()
         {
             this.cacheControllerUtil.ClearCache("nosuchcache");
@@ -123,7 +124,7 @@ namespace PubComp.Caching.Core.UnitTests
         }
 
         [TestMethod]
-        [ExpectedException(typeof(CacheException))]
+        [ExpectedException(typeof(CacheClearException))]
         public void TestRegisterCacheWithFalse_ClearCache()
         {
             this.cacheControllerUtil.RegisterCache("cache2", false);
@@ -132,14 +133,14 @@ namespace PubComp.Caching.Core.UnitTests
         }
 
         [TestMethod]
-        [ExpectedException(typeof(CacheException))]
+        [ExpectedException(typeof(CacheClearException))]
         public void TestRegisterCache_EmptyName()
         {
             this.cacheControllerUtil.RegisterCache(string.Empty, true);
         }
 
         [TestMethod]
-        [ExpectedException(typeof(CacheException))]
+        [ExpectedException(typeof(CacheClearException))]
         public void TestRegisterCache_NullName()
         {
             this.cacheControllerUtil.RegisterCache(null, true);
@@ -152,28 +153,28 @@ namespace PubComp.Caching.Core.UnitTests
         }
 
         [TestMethod]
-        [ExpectedException(typeof(CacheException))]
+        [ExpectedException(typeof(CacheClearException))]
         public void TestRegisterCacheItem_NullCacheName()
         {
             this.cacheControllerUtil.RegisterCacheItem<SubClass1>(null, "keyA");
         }
 
         [TestMethod]
-        [ExpectedException(typeof(CacheException))]
+        [ExpectedException(typeof(CacheClearException))]
         public void TestRegisterCacheItem_EmptyCacheName()
         {
             this.cacheControllerUtil.RegisterCacheItem<SubClass1>(string.Empty, "keyA");
         }
 
         [TestMethod]
-        [ExpectedException(typeof(CacheException))]
+        [ExpectedException(typeof(CacheClearException))]
         public void TestRegisterCacheItem_NullItemKey()
         {
             this.cacheControllerUtil.RegisterCacheItem<SubClass1>("cache1", null);
         }
 
         [TestMethod]
-        [ExpectedException(typeof(CacheException))]
+        [ExpectedException(typeof(CacheClearException))]
         public void TestRegisterCacheItem_EmptyItemKey()
         {
             this.cacheControllerUtil.RegisterCacheItem<SubClass1>("cache1", string.Empty);
@@ -191,7 +192,7 @@ namespace PubComp.Caching.Core.UnitTests
         }
 
         [TestMethod]
-        [ExpectedException(typeof(CacheException))]
+        [ExpectedException(typeof(CacheClearException))]
         public void TestRegisterCacheWithFalse_RegisterCacheItems_ClearCache()
         {
             this.cacheControllerUtil.RegisterCache("cache2", false);
@@ -203,7 +204,7 @@ namespace PubComp.Caching.Core.UnitTests
         }
 
         [TestMethod]
-        [ExpectedException(typeof(CacheException))]
+        [ExpectedException(typeof(CacheClearException))]
         public void TestCacheRegisterCacheItems_ClearCache()
         {
             this.cacheControllerUtil.RegisterCacheItem(
@@ -236,21 +237,21 @@ namespace PubComp.Caching.Core.UnitTests
         }
 
         [TestMethod]
-        [ExpectedException(typeof(CacheException))]
+        [ExpectedException(typeof(CacheClearException))]
         public void TestRegisterCacheWithTrue_RegisterCacheItems_ClearCacheItem_WrongCacheName()
         {
             this.cacheControllerUtil.ClearCacheItem("nosuchcache", "keyA");
         }
 
         [TestMethod]
-        [ExpectedException(typeof(CacheException))]
+        [ExpectedException(typeof(CacheClearException))]
         public void TestRegisterCacheWithTrue_RegisterCacheItems_ClearCacheItem_EmptyKey()
         {
             this.cacheControllerUtil.ClearCacheItem("cache1", string.Empty);
         }
 
         [TestMethod]
-        [ExpectedException(typeof(CacheException))]
+        [ExpectedException(typeof(CacheClearException))]
         public void TestRegisterCacheWithTrue_RegisterCacheItems_ClearCacheItem_NullKey()
         {
             this.cacheControllerUtil.ClearCacheItem("cache1", null);
@@ -303,35 +304,35 @@ namespace PubComp.Caching.Core.UnitTests
         }
 
         [TestMethod]
-        [ExpectedException(typeof(CacheException))]
+        [ExpectedException(typeof(CacheClearException))]
         public void TestCacheRefreshItem_NoSuchCache()
         {
             this.cacheControllerUtil.RefreshCacheItem("nosuchcache", "keyB");
         }
 
         [TestMethod]
-        [ExpectedException(typeof(CacheException))]
+        [ExpectedException(typeof(CacheClearException))]
         public void TestCacheRefreshItem_NullCacheName()
         {
             this.cacheControllerUtil.RefreshCacheItem(null, "keyB");
         }
 
         [TestMethod]
-        [ExpectedException(typeof(CacheException))]
+        [ExpectedException(typeof(CacheClearException))]
         public void TestCacheRefreshItem_EmptyCacheName()
         {
             this.cacheControllerUtil.RefreshCacheItem(string.Empty, "keyB");
         }
 
         [TestMethod]
-        [ExpectedException(typeof(CacheException))]
+        [ExpectedException(typeof(CacheClearException))]
         public void TestCacheRefreshItem_NullKey()
         {
             this.cacheControllerUtil.RefreshCacheItem("cache1", null);
         }
 
         [TestMethod]
-        [ExpectedException(typeof(CacheException))]
+        [ExpectedException(typeof(CacheClearException))]
         public void TestCacheRefreshItem_EmptyKey()
         {
             this.cacheControllerUtil.RefreshCacheItem(string.Empty, null);